Add substring-after.
Right now it returns a function instead of doing the work immediately.
It seems like this is more efficient (don't have to pass the delim over
and over), but (a) it might be confusing for no reason, and (b) I don't
know that it's actually more efficient. Some follow-up needed here.